Cleaning of Strings
I need to clean a few more columns. I'll do PRs for each. A future PR
will address geocoding 'owner_mailing_address'.

This PR includes a pickle of a dictionary I created, composed of geocoded
results as its value and they as a zipcode. Before geocoding, it checks
for its absence in the dictionary. How I split the string did not cover
for all cases–those errors are indexable, however, since we're dealing
with a dictionary and a dataframe. Inspite of being time consuming, the
geocoder seems to be fairly accurate, apart from the occasional
abbreviated and unrecognized street name.