Added a virtualenv for Python, but…
Poetry has one too? So…
Jersey City's jupyter notebook is now linking to a github url instead of
the path to the folder on my other computer. That link is the .csv for
the raw_df.

Progress on cleaning the owner_name column. I am using a dictionaries to
keep track of filtered groups.